+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Added `id_token` to `AccessCredentials`
 
+* Migrated to Dart 2 `BigInt`.
+
 ## 0.2.3+6
 
 - Fix async issue in oauth2 flow implementation